refactor PaymentMethod list
